% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{c0}
\alias{c0}
\title{Taiwan Elevation Data (c0-c8)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 2843 rows and 4 columns.
}
\usage{
c0
}
\description{
A series of spatial feature (sf) polygon objects representing different 
elevation bands (c0 represents elevation band 0-250m).
}
\keyword{datasets}
