% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{c1}
\alias{c1}
\title{Taiwan Elevation Data (Band 2: 250-500m)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1990 rows and 4 columns.
}
\usage{
c1
}
\description{
A spatial feature (sf) polygon object representing elevation band 250-500m.
}
\keyword{datasets}
