% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{c2}
\alias{c2}
\title{Taiwan Elevation Data (Band 3: 500-1000m)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 533 rows and 4 columns.
}
\usage{
c2
}
\description{
A spatial feature (sf) polygon object representing elevation band 500-1000m.
}
\keyword{datasets}
