% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{c3}
\alias{c3}
\title{Taiwan Elevation Data (Band 4: 1000-1500m)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 329 rows and 4 columns.
}
\usage{
c3
}
\description{
A spatial feature (sf) polygon object representing elevation band 1000-1500m.
}
\keyword{datasets}
