% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{c4}
\alias{c4}
\title{Taiwan Elevation Data (Band 5: 1500-2000m)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 244 rows and 4 columns.
}
\usage{
c4
}
\description{
A spatial feature (sf) polygon object representing elevation band 1500-2000m.
}
\keyword{datasets}
