% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{c5}
\alias{c5}
\title{Taiwan Elevation Data (Band 2: 2000-2500m)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 187 rows and 4 columns.
}
\usage{
c5
}
\description{
A spatial feature (sf) polygon object representing elevation band 2000-2500m.
}
\keyword{datasets}
