% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{c6}
\alias{c6}
\title{Taiwan Elevation Data (Band 2: 2500-3000m)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 131 rows and 4 columns.
}
\usage{
c6
}
\description{
A spatial feature (sf) polygon object representing elevation band 2500-3000m.
}
\keyword{datasets}
