% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{c7}
\alias{c7}
\title{Taiwan Elevation Data (Band 2: 3000-3500m)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 101 rows and 4 columns.
}
\usage{
c7
}
\description{
A spatial feature (sf) polygon object representing elevation band 3000-3500m.
}
\keyword{datasets}
