% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{c8}
\alias{c8}
\title{Taiwan Elevation Data (Band 2: 3500-4000m)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 27 rows and 4 columns.
}
\usage{
c8
}
\description{
A spatial feature (sf) polygon object representing elevation band 3500-4000m.
}
\keyword{datasets}
