% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daly_B01.R
\docType{data}
\name{daly_B01}
\alias{daly_B01}
\title{Demo dataset: DALY estimates of cancers (level-3)}
\format{A data frame with 352 rows and 7 variables:
\describe{
  \item{sex}{0 for female and 1 for male.}
  \item{town}{town code, MOI format.}
  \item{year}{2015 only}
  \item{cause_l3}{level-3 cause code}
  \item{daly_count}{DALY estimates (person-year)}
  \item{yld_count}{YLD estimates (person-year)}
  \item{yll_count}{YLL estimates (person-year)}
  \item{population}{population size}

}}
\source{
TBDC
}
\usage{
daly_B01
}
\description{
A dataset containing the DALY, YLD and YLL of level-3 cancers (Taiwan 2015).
}
\keyword{datasets}
