% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_cause.R
\docType{data}
\name{dict_cause}
\alias{dict_cause}
\title{Reference dataset: dictionary for cause outline and name}
\format{A data frame with 14256 rows and 5 variables:
\describe{
  \item{cause_id}{city code, MOI format }
  \item{cause_outline}{}
  \item{acause}{}
  \item{cause_name}{}
  \item{cause_short}{}
  \item{level}{}
  \item{cause_name_ch}{}
}}
\source{
TBDC
}
\usage{
dict_cause
}
\description{
Reference dataset: dictionary for cause outline and name
}
\keyword{datasets}
