% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{elevprof}
\alias{elevprof}
\title{Elevation Profile Data}
\format{
An object of class \code{data.frame} with 9495 rows and 2 columns.
}
\usage{
elevprof
}
\description{
A data frame containing pre-calculated latitude-elevation profile data.
}
\keyword{datasets}
