% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_map.R
\name{export_map}
\alias{export_map}
\title{Export a map}
\usage{
export_map(fname = NULL, map = NULL)
}
\arguments{
\item{fname}{file name.}

\item{map}{a map generated by \code{twmap_static()} or a ggplot object.}
}
\value{
NULL.
}
\description{
Export a map
}
