% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{river}
\alias{river}
\title{River Data}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 411 rows and 9 columns.
}
\usage{
river
}
\description{
A spatial feature (sf) line object representing major rivers.
}
\keyword{datasets}
