% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{sample_spts.ly}
\alias{sample_spts.ly}
\title{Species Distribution Sample Points for Lanyu (WGS84)}
\format{
A data frame with 31 rows and 3 variables:
\describe{
  \item{x}{WGS84 Longitude (decimal degrees).}
  \item{y}{WGS84 Latitude (decimal degrees).}
  \item{z}{Elevation in meters.}
}
}
\source{
Taiwan Biodiversity Network (10 September 2020) TBN Occurrence Download.
}
\usage{
data(sample_spts.ly)
}
\description{
A sample data frame containing *Maesa lanyuensis* Yuen P.Yang distribution records 
intended for testing the \code{distrmap} function. This data is provided 
for demonstration and testing purposes.
}
\keyword{datasets}
