% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{sample_spts.tw}
\alias{sample_spts.tw}
\title{Species Distribution Sample Points for Taiwan (WGS84)}
\format{
A data frame with 30 rows and 3 variables:
\describe{
  \item{x}{WGS84 Longitude (decimal degrees).}
  \item{y}{WGS84 Latitude (decimal degrees).}
  \item{z}{Elevation in meters.}
}
}
\source{
Simulated data.
}
\usage{
data(sample_spts.tw)
}
\description{
A sample data frame containing fictitious species distribution records 
intended for testing the \code{distrmap.tw} function. This data is provided 
for demonstration and testing purposes.
}
\keyword{datasets}
