% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_pop.R
\docType{data}
\name{stand_pop}
\alias{stand_pop}
\title{Reference dataset: Standard population for age-standardization}
\format{A data frame with 23 rows and 6 variables:
\describe{
  \item{age1}{ age group in categorization 1}
  \item{percent1}{ percentage in population of age group in categorization 1}
  \item{age2}{ age group in categorization 2}
  \item{percent2}{ percentage in population of age group in categorization 2}
  \item{age3}{ age group in categorization 3}
  \item{percent3}{ percentage in population of age group in categorization 3}
}}
\source{
TBDC
}
\usage{
stand_pop
}
\description{
Reference dataset: Standard population for age-standardization
}
\keyword{datasets}
