% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_county.R
\docType{data}
\name{tw_county}
\alias{tw_county}
\title{Taiwan county map produced by NLSC(EPSG:3823, TWD97)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 22 rows and 5 columns.
}
\source{
\url{https://data.gov.tw/dataset/7442}
}
\usage{
tw_county
}
\description{
Last update: 2019-11-21
}
\keyword{datasets}
