% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_county_adj.R
\docType{data}
\name{tw_county_adj}
\alias{tw_county_adj}
\title{Taiwan county map produced by NLSC(EPSG:3824, TWD97)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 22 rows and 6 columns.
}
\source{
\url{https://data.gov.tw/dataset/7442}
}
\usage{
tw_county_adj
}
\description{
Last update: 2019-11-21
}
\details{
Adjusted Part:
- remove uninhabited island
- reduce the distance between other island and Taiwan
- reduce the size of geometry file
}
\keyword{datasets}
