% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_town.R
\docType{data}
\name{tw_town}
\alias{tw_town}
\title{Taiwan town map produced by NLSC(EPSG:3823, TWD97)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 368 rows and 8 columns.
}
\source{
\url{https://data.gov.tw/dataset/7442}
}
\usage{
tw_town
}
\description{
Taiwan town map produced by NLSC(EPSG:3823, TWD97)
}
\keyword{datasets}
