% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_town_adj.R
\docType{data}
\name{tw_town_adj}
\alias{tw_town_adj}
\title{Taiwan town map produced by NLSC(EPSG:3824, TWD97)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 368 rows and 9 columns.
}
\source{
\url{https://data.gov.tw/dataset/7441}
}
\usage{
tw_town_adj
}
\description{
Last update: 2020-03-24
}
\details{
Adjusted Part:
- remove uninhabited island
- reduce the distance between other island and Taiwan
- reduce the size of geometry file
}
\keyword{datasets}
