% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_village.R
\docType{data}
\name{tw_village}
\alias{tw_village}
\title{Taiwan village map produced by NLSC(EPSG:3824, TWD97)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 7965 rows and 11 columns.
}
\source{
\url{https://data.gov.tw/dataset/7438}
}
\usage{
tw_village
}
\description{
Last update: 2019-04-23
}
\keyword{datasets}
