% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{twbound}
\alias{twbound}
\title{Taiwan Boundary}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 2606 rows and 4 columns.
}
\usage{
twbound
}
\description{
A spatial feature (sf) polygon object representing the main island boundary.
}
\keyword{datasets}
