\name{twmap-package}
\alias{twmap-package}
\alias{twmap}
\docType{package}
\title{Draw species plots and elevation in Taiwan}
\description{
It's lazy-man-drawing Taiwan map and species data program with classified elevation and rivers.
}
\details{
\tabular{ll}{
Package: \tab twmap\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-05-25\cr
License: \tab FreeBSD License\cr
LazyLoad: \tab yes\cr
}
drawDTMSP()
drawtwmap()
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Lin, Cheng-Tao
Maintainer: Lin, Cheng-Tao <mutolisp@gmail.com>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ graphics, map, Taiwan }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
