% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWmap.R
\docType{package}
\name{TWmap}
\alias{TWmap}
\alias{TWmap-package}
\title{TWmap: A package for making Taiwan maps by adminstrative divisions.}
\description{
TWmap: A package for making Taiwan maps by adminstrative divisions.
}
\section{Introduction}{

The  functions ...
}

\keyword{internal}
