% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWpop_city.R
\docType{data}
\name{TWpop_city}
\alias{TWpop_city}
\title{Reference dataset: Taiwan population by age, sex, year, and city}
\format{A data frame with 14256 rows and 5 variables:
\describe{
  \item{city}{city code, MOI format }
  \item{age}{}
  \item{sex}{}
  \item{year}{}
  \item{population}{}
}}
\source{
TBDC
}
\usage{
TWpop_city
}
\description{
Reference dataset: Taiwan population by age, sex, year, and city
}
\keyword{datasets}
