% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWpop_region.R
\docType{data}
\name{TWpop_region}
\alias{TWpop_region}
\title{Reference dataset: Taiwan population by age, sex, year, and region}
\format{A data frame with 14256 rows and 5 variables:
\describe{
  \item{region}{region code }
  \item{age}{}
  \item{sex}{}
  \item{year}{}
  \item{population}{}
}}
\source{
TBDC
}
\usage{
TWpop_region
}
\description{
Reference dataset: Taiwan population by age, sex, year, and region
}
\keyword{datasets}
