% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWpop_town.R
\docType{data}
\name{TWpop_town}
\alias{TWpop_town}
\title{Reference dataset: Taiwan population by age, sex, year, and townships}
\format{A data frame with 238464 rows and 5 variables:
\describe{
  \item{town}{town code, MOI format }
  \item{age}{}
  \item{sex}{}
  \item{year}{}
  \item{population}{}
}}
\source{
TBDC
}
\usage{
TWpop_town
}
\description{
Reference dataset: Taiwan population by age, sex, year, and townships
}
\keyword{datasets}
