% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{twsp}
\alias{twsp}
\title{Species Distribution Data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 10 rows and 2 columns.
}
\usage{
twsp
}
\description{
A spatial matrix containing species points.
}
\keyword{datasets}
