% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{vertplots}
\alias{vertplots}
\title{Vertical Plot Coordinates}
\format{
An object of class \code{data.frame} with 20 rows and 2 columns.
}
\usage{
vertplots
}
\description{
A data frame containing coordinates for vertical plots.
}
\keyword{datasets}
