% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attri_death_pm25.R
\docType{data}
\name{attri_death_pm25}
\alias{attri_death_pm25}
\title{Demo dataset: Deaths attributable to PM.2.5 (Taiwan 2013)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 352 rows and 7 columns.}
\source{
Nicholas
}
\usage{
attri_death_pm25
}
\description{
A dataset containing the number of attributable deaths to PM2.5 exposure by townships. The values are simulated
from the original dataset provided by Nicholas.
}
\keyword{datasets}
